requirements.txt
python-ldap==3.3.1
python-arango==7.2.0